refactor: 고객 관련 controller customer 패키지로 이동
