+++ src/main/java/org/programmers/springorder/VoucherApplication.java
@@ -1,8 +1,8 @@
 package org.programmers.springorder;
 
 import org.programmers.springorder.console.Console;
-import org.programmers.springorder.controller.CustomerController;
 import org.programmers.springorder.controller.WalletController;
+import org.programmers.springorder.controller.customer.CustomerController;
 import org.programmers.springorder.controller.voucher.VoucherController;
 import org.programmers.springorder.utils.ExceptionHandler;
 import org.programmers.springorder.utils.MenuType;
